+++ src/s1/components/app.jsx
@@ -1,6 +1,7 @@
 import { h, Component } from 'preact';
 import BannerStacks from './banner/bannerStacks';
 import BannerVendors from './banner/bannerVendors';
+import BannerSlim from './banner/bannerSlim';
 
 import { CONSENT_SCREENS } from '../constants';
 
@@ -47,10 +48,18 @@ export default class App extends Component {
 
 		return (
 			<div class={style.gdpr}>
-				{!consentScreen ||
+				{
+					!consentScreen ||
 					(consentScreen === CONSENT_SCREENS.STACKS_LAYER1 && (
 						<BannerStacks store={store} isShowing={shouldShowModal && tcModel} />
-					))}
+					))
+				}
+
+				{  
+					(consentScreen === CONSENT_SCREENS.SLIM_LAYER0 && (
+						<BannerSlim store={store} isShowing={shouldShowModal && tcModel} />
+					))
+				}
 
 				{consentScreen === CONSENT_SCREENS.VENDORS_LAYER3 && (
 					<BannerVendors store={store} isShowing={shouldShowModal && tcModel} />