+++ src/s1/lib/config.js
@@ -18,6 +18,8 @@ export const theme = {
 	// secondaryColor: '#869cc0',
 	// featuresColor: '#d0d3d7',
 	// maxWidthModal: '500px',
+	isFullWidth: false,
+	shouldShowDropShadow: true,
 	shouldAutoResizeModal: true,
 	maxHeightModal: '45vh',
 };
@@ -32,6 +34,8 @@ export const config = {
 	cookieDomain: '',
 	ccpaApplies: false,
 	experimentId: 'control',
+	isSlimMode: false,
+	shouldShowCloseX: false,
 	gdprApplies: false,
 	gdprConsentUrlParam,
 	isServiceSpecific: true, // whether or not this cmp is configured to be service specific